requirements.txt
numpy
pandas
matplotlib
seaborn
nltk
wordcloud
scikit-learn
pillow
tk
